Refactor Restart.py -> restart.py
