+++ medusa/server/api/v1/core.py
@@ -54,7 +54,7 @@ from ....search.queue import BacklogQueueItem, ForcedSearchQueueItem
 from ....show.ComingEpisodes import ComingEpisodes
 from ....show.History import History
 from ....show.Show import Show
-from ....system.Restart import Restart
+from ....system.restart import Restart
 from ....system.shutdown import Shutdown
 from ....version_checker import CheckVersion
