+++ pytest.ini
@@ -209,7 +209,7 @@ flake8-ignore =
     medusa/show_name_helpers.py D100 D200 D202 D205 D400 D401 N802 N803 N806
     medusa/show_queue.py D100 D101 D102 D205 D400 D401 E501 N802 N803 N806
     medusa/show_updater.py D100 D101 D102 D105
-    medusa/system/Restart.py D100 D101 D102
+    medusa/system/restart.py D100 D101 D102
     medusa/system/shutdown.py D100 D101 D102
     medusa/system/__init__.py D104
     medusa/tagger/__init__.py D104
